move reverify tests to reverify_test.go
